requirements.txt
cisco-gnmi==1.0.11
PyYAML==5.3.1
